backend/ml-services/visual-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ML Libraries
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
Pillow==10.1.0
scikit-image==0.22.0

# Web Rendering
playwright==1.40.0
selenium==4.15.2
beautifulsoup4==4.12.2

# Image Processing
imagehash==4.3.1
numpy==1.24.3
scikit-learn==1.3.2

# AWS
boto3==1.29.7

# Utilities
python-dotenv==1.0.0
redis==5.0.1
httpx==0.25.2
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
